Fix bugs encountered with travis-ci
Bugs in the library that halted compiling that were discovered as
part of running travis-ci have been address.

In particular, an issue with using __has_cxx_attribute was failing
on older versions of GCC, since a macro test condition for this
feature was not being short-circuited correctly.
To allow this to work on older compilers, this was removed and
replaced with direct use of GCC's __attribute__ syntax instead.