Fix travis conan deployment
The .travis.yml file had the trigger for conan deployment
being a tag starting with a 'v', and being on 'master'. These
two requirements are actually mutually exclusive, and thus this
trigger never fired automatically.

To fix this, the trigger has been changed to only fire if a tag
is created.